solve part 1 for day 16
This has been my worst effort yet.  This is a math puzzle, and here I am
writing code that iteratively tries guesses.  I didn't even work out a
sensible bound for such guesses!  I just chose 1,000 because it's a big
number and went with that.  This 'solution' only outputs a guess too!
Without the external check, there was no way to know if the result was
correct.

Terrible, terrible.  And this is only part 1 of day 17.  And today is
day 22.