a fresh-ish start for my attempt at day 17 part 2
It seems my part 1 solution worked despite being wrong in places.

Also: there is indeed a math solution to part 1, and it is as simple as
the triangular number formula - though I can't see how.

I'd guess there's a math solution to part 2 too, but I'm going to
continue with the stupid iterative approach.