solve day 18 part 1
I had misread the spec.  Unfortunately my implementation worked for most
test cases anyway, so I became convinced that I had a bug in my
implementation, but the problem was deeper than that.  Anyway, working
now.

I would like to reimplement this in a smarter way, and I'll get around
to that some other time.