Solve day 14 part 2
Yeah.  Ok, so the code here doesn't exactly solve the problem, but it
gets pretty close.

The state of the platform starts to repeat from the 136th iteration
onwards, in a cycle of 11 states.

Odd powers of ten have a remainder of 10 when divided by 11, so
1,000,000,000 (10^9) is at the same point of the cycle as 1,000 (10^3).
Running the cycle function 1,000 times took only 30s or so ... so I
couldn't be bothered writing code to find the answer when it's right
there.